Requested Change + QoL adjustments
Suggested change, also noticed execute using BTU instead of KJ (we should stick to SI units).

For a QoL it would be best if we could run this without having to reopen and close Midas each time.

Finally, as a suggestion most people are on a later version of Midas now so we should think about updating the 8.8.1 to 8.8.4.